Fixes both:
[3041096]  First linein on read-only file returns empty line
[3041121]  Hostemu (execio) can't read read-only file in windows 7

On an implicit open, we first try to open as read write.  This fails
for a read only file.  We then try to open in read only mode.  This
succeeds, but since we didn't clear errors from the first attempt,
the linein() which caused the implicit open fails.  Since the file is
open, succeeding lineins work.


git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/main/trunk@6090 0b6cbdbe-3aab-466e-b73a-abd511dda0a2